Add an OnFire ability to the player.
This commit adds an OnFire ability to the player that activates if you bump into a tag set in the tagName property in the PlayerOnFire script and that lets you set yourself on fire for a set amount of time with a key you set. In the example scene, that's the F key and the amount of time that the Player waits is 2 seconds.

I also commented quite a bit of code and removed the hard dependencies of Animator from the Player script and now the Player script checks if you have the Animator and PlayerOnFire scripts attached to the GameObject, and if not, it will set internal bools that wont let the functions work.