The TileOutOfBounds script now restores the original bounds.
This script now restores the original bounds in the camera view when you enter the original tilemap bounds.

I also changed it to use the Contains method instead of checking positions manually with math as this is more accurate, and it will work if you extend bounds from the right side instead of the left, and it just works better.