Finish the IKEA level.
This commit finishes the IKEA level by adding a lot of stuff to it. First, theres a trick path right at the beginning in the front where if you go to the end, it flicks you back 30 meters to the starting area. I also extended the level quite a bit, adding another trick path where its a big trick path above with barriers that trap you in until you get to the end.