Change the FlamingAd script to teleport and disable movement.
This commit changes the FlamingAd script from being a script where you click multiple times and run methods you specify in an event, to being a script that teleports you to a location and disables your players movement when you run into a trigger. I obsoleted the previous code so you can still use it, you just have to press a checkbox and I coded some fancy Unity Inspector UI stuff where it disables the fields when that boolean is unchecked.