Improve the Player Script's Inspector.
This commit adds a new PlayerEditor script that acts as a Custom Inspector for the Player script that uses the base default inspector up to the groundTag bool, then it uses code in PlayerEditor to draw custom fields that disable when a condition is met. The groundTag field is disabled when checkForGround is false and the Animation and OnFire properties are disabled unless you attach the component to the script. The properties also have a special message telling you to add the component for that feature to work.