Add a currency system.
This commit adds a currency system to the Game that works by using the GameManager. The GameManager has methods you can use to modify the money stuff and when you modify the money, it fires a UnityEvent that can either return a String or a Float according to what Return Type you set.

This commit also adds a GoldStorage script that generates money according to an amount and a time between generation and it uses a coroutine to wait after it adds money.