Make a fully fledged Storage system.
This commit makes a fully fledged storage system with a Custom Inspector. The Inspector has a preset system where you can either set the storage script to use a Gold Storage preset or a Elixir preset and theres a button that opens a custom Editor Window that lets you modify the preset settings so you can change the preset values to set different values if you don't feel like the defaults are what you want.