I give up See you all 04/10/23
I'm at a loss of what to do on several scripts. Ill Create some scripts for mob spawning tomorow but until some other scripts and key features are fixed I can't progress.  I might just scrap this for parts and have it be something you just sit through like an unscipable ad but this is almost a whole ass game and I'm realizing how unrealistic this was.