Add a Drag and Drop script.
This commit adds a Drag and Drop script that lets you click on the object with the script and a collider and it lets you drag the object with your mouse. If you have a Rigidbody2D on the object, it will turn the velocity to 0 if you let go of the object.