Create a Pause Menu and WIP Settings Menu.
This commit creates a Pause Menu that works by Instantiate a prefab with a Canvas with a Pause Menu. It lets you either resume the game, go to the Work in Progress Settings Menu which is also another prefab, and you can quit back to the Title Screen.

The Settings Menu will let you change the volume and the resolution but it currently doesn't have any of that logic implemented so when you go to the settings menu, it will just have a broken menu.