Add Incorrect Object Position to the Sockets script.
This commit adds the Incorrect Object Position feature to the Sockets script letting you teleport the Object you put into the Socket if its not the Object you set in the Correct GameObject field. By default, if you have the Incorrect Object Position to <0, 0>, no teleporting will happen. You can override that with a toggle.