Make a Timer.
This commit makes a Timer that counts down with an animation and it shows how many seconds are left in the clock. When the timer is out of time, it will fire off a Unity Event.

This timer is manually animated by using a for loop and it waits around (1 / 17 fps) seconds per loop.

I also unhooked the Collider2D dependency from the ActivateDeactivate script so you can put it on container GameObjects without requiring a Collider2D and I made the Activate and Deactivate methods public.