Improve the Animation experience on the Player script.
There are alot of changes with this commit.

First, I added jumping support so you can choose a parameter and it will go to true if you jump.

Second, I programmed the editor to read the Parameters on the Controller on the Animator and let you choose between then and it will save into the file.