Improve the Debug Menu and make some abstractions.
This commit improves the Debug Menu by making a script called DebugSupport which you attach to your Ad Objects and you put where you want the Player to teleport to and the script will handle all of the settings the Active Scene in the GameManager. The GameManager Ad list will warn you if you don't have a DebugSupport script on the GameObject you put in there.

I also added some internal abstractions for Editor stuff.