Convert all current Ads to Prefabs.
This commit converts all current Ads to prefabs to make them easier to manage and hopefully will give less conflicts.

I also changed DebugSupport to give you an option to Enable the Main Player on the Scene to the Player doesn't go on scenes that arent designed for the Main Player.