Bring Back the Two Tongue ad and fix the sockets.
This commit brings back the Two Tongue ad into a prefab from an old version before it got erased from the main scene. I put white square placeholders that have a socket script and populated the sockets script with some relating to the sockets and correct GameObjects.

This commit also fixes the Player script not doing anything with the Jumping animation parameter.