Add Sound Effects to the Player Controller.
This commit adds Sound Effects to the Player Controller. This means you can hook the Walking and Jumping actions to a Sound Effect.

I also fixed the DebugMenu by making a MainPlayer script so when the MainPlayer starts with that script attached, it tells the GameManager that it is the Main Player.