Create a portable HealthSystem.
This commit creates an entire brand new Health script which is a component that contains data about Health. I extracted the fields that controlled max health and the WhenYouDie event out from the GameManager and into the script file so you can set those per Health Component.

I re-engineered the HealthBar script to accept Health Components instead of defaulting to using the GameManager's health system, but theres an toggle you can click to let it use the GameManagers HealthSystem which has a type of Global Health system that you can have by attaching a Health Component to the GameManager and it will re-export it out so you can access the GameManager's Health component from a GameManager object without having to GetComponent it.