Create a Manual Coordinates Camera system.
This commit creates a Manual Coordinates Camera system. This means you can create a ManualCoordinates asset file and you can input minimum and maximum coordinates and the camera will clamp to those coordinates. You need to add a ManualCoordinates script to the main GameObject of the Ad and provide the asset file inside the inspector and the script should provide the camera with the coordinates and the Cam script should do the rest.

This also gave me the opportunity to improve the Cam script by making it more modular. I made a ICameraClamping interface that scripts can implement so its relatively easy to add new Camera Clamping interfaces to the script. You just need to add a new value to the Enum, and make a class implementing the logic you want.

For example, I moved the Tilemap clamping to its own script, and implemented the methods in the ICameraClamping interface so you can have the choice between clamping to a tilemap, or using ManualCoordinates files.