Polishing up the Game.
This commit polishes various things about the game. Like how I added the Manual Coordinates system to Stage-1 so the Camera clamps perfectly to Stage-1.

Theres also a script called CamFollowObject which runs when you enable the GameObject which attaches the Object you give to it as the Cam's FollowedObject and when the GameObject gets disabled, it un-attaches it.

I also added a way to control the Cam's clamping. The ClampControl script. Attach this script to a GameObject, and it will modify the camera for you, and disable clamping when the GameObject gets disabled. The Inspector is a replica of the Cam's Clamping Properties, and these properties will only apply when the GameObject is enabled. So this means the Camera, by default, has no followed object, and clamping disabled. So you have to use scripts to provide the Camera with the information and control it.