Polish up the camera a bit more.
This commit polishes up the camera a bit more. For example, the camera doesn't do that smoothing motion when you switch Ad's. It's now programmed to be instant, and then it will smooth.

I also added a option to the Settings to let you disable the Camera Smoothing, so you can disable it if you don't like Camera Smoothing.