Add the beginnings of a PlayerWallClimbing script.
This commit adds the beginnings of a PlayerWallClimbing script, which should let you climb up a wall when you bump into a wall.

Right now I implemented a Raycast that points out at the direction the player is facing. Right now, the distance of the Raycast is 1 meter, but I'm pretty sure I should make it shorter.